Changed weight to be a float in userPreference. Added two weight manipulation functions in the calcs. One of these changes the weight of a specific preference. The other tells you how much weight a preference has to work with on the pie chart.

+++ College Rank/PreferenceManager.m
@@ -62,12 +62,12 @@
 }
 
 
--(void) addUserPref: (Preference*) pref withWeight: (int) weight
+-(void) addUserPref: (Preference*) pref withWeight: (float) weight
 {
     [self.userPrefs addObject:[[UserPreference alloc] initWithPreference:pref andWeight:weight]];
 }
 
--(void) addUserPref: (Preference*) pref withWeight: (int) weight andPrefVal: (int) prefVal
+-(void) addUserPref: (Preference*) pref withWeight: (float) weight andPrefVal: (int) prefVal
 {
     [self.userPrefs addObject:[[UserPreference alloc] initWithPreference:pref andWeight:weight andPrefVal:prefVal]];
 }
